requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pandas==2.2.2
openpyxl==3.1.2
pdfplumber==0.10.3
python-docx==1.1.0
jinja2==3.1.4
